docker/iota-private-network/scripts/requirements.txt
docker==7.1.0
PyYAML==6.0.1